requirements.txt
pytest==6.2.5
selenium==3.141.0
webdriver-manager==3.4.2
pyhamcrest==2.0.2
pyyaml==5.4.1